move emote related code to emote manager
