+++ src/plugin-hyperfy/utils.ts
@@ -1,0 +1,7 @@
+export async function hashFileBuffer(buffer: Buffer): Promise<string> {
+    const hashBuf = await crypto.subtle.digest('SHA-256', buffer)
+    const hash = Array.from(new Uint8Array(hashBuf))
+      .map(b => b.toString(16).padStart(2, '0'))
+      .join('')
+    return hash
+}
